[aws] Move container implementation to a separate class
